+++ enclone_visual/src/enclone_client.rs
@@ -28,13 +28,13 @@
 // The special argument SERVER_DEBUG causes the server to print debuggin information.  However
 // you will only see this if you run the server locally using enclone VIS.
 
+use crate::launch_gui;
 use crate::proto::{analyzer_client::AnalyzerClient, ClonotypeRequest, EncloneRequest};
 use crate::update_restart::*;
 use crate::*;
 use enclone_core::parse_bsv;
 use enclone_core::prepare_for_apocalypse::*;
 use enclone_core::REMOTE_HOST;
-use gui::launch_gui;
 use io_utils::*;
 use itertools::Itertools;
 use libc::atexit;
